requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.1
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==44.0.3
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
document==1.0
durationpy==0.9
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.9.0.post1
fastapi==0.115.5
ffmpy==0.5.0
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
google-auth==2.38.0
google_search_results==2.4.2
googleapis-common-protos==1.67.0
gradio==5.16.0
gradio_client==1.7.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.7.1
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==32.0.0
langchain==0.3.22
langchain-anthropic==0.3.10
langchain-chroma==0.2.2
langchain-community==0.3.20
langchain-core==0.3.49
langchain-experimental==0.3.4
langchain-mongodb==0.2.0
langchain-openai==0.2.9
langchain-postgres==0.0.12
langchain-text-splitters==0.3.7
langgraph==0.2.53
langgraph-checkpoint==2.0.25
langgraph-checkpoint-mongodb==0.1.3
langgraph-checkpoint-redis==0.0.5
langgraph-sdk==0.1.36
langsmith==0.1.145
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mmh3==5.1.0
monotonic==1.6
motor==3.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.55.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.11
ormsgpack==1.9.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pgvector==0.2.5
pillow==11.1.0
platformdirs==4.3.8
ply==3.11
posthog==3.13.0
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==5.29.3
psutil==7.0.0
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.3
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-ulid==3.0.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
redisvl==0.5.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.2
selenium==4.32.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tavily-python==0.5.3
tenacity==9.0.0
tiktoken==0.8.0
together==1.5.7
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.48.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.0
zipp==3.21.0
